backend/requirements.txt
fastapi
uvicorn
torch
sentence-transformers
faiss-cpu
whoosh
transformers
pandas
scikit-learn 